feat : apply suspense for data-fetching
