+++ src/pages/leagueoflegends/CardDetailFetcher.tsx
@@ -3,7 +3,6 @@ import { useParams } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import { useErrorBoundary } from 'react-error-boundary';
 
-import { useGetData } from 'useGetData';
 import { cardActions } from 'store/card-slice';
 import CardDetailContainer from './CardDetailContainer';
 
@@ -16,10 +15,6 @@ const CardDetailFetcher = () => {
 
   const params = useParams();
   const { id: cardId } = params;
-
-  if (cardId) {
-    const response: any = useGetData(`/api/lol/boards/${cardId}`);
-  }
 
   return <CardDetailContainer />;
 };
