+++ src/pages/leagueoflegends/CardListFetcher.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
-import { defaultAxios } from 'apis/utils';
 import { useDispatch } from 'react-redux';
-import { useErrorBoundary } from 'react-error-boundary';
 
 import { cardActions } from 'store/card-slice';
+import useGetData from 'useGetData';
 
 interface CardListFetcherProps {
   fetcherProps: {
@@ -19,26 +18,19 @@ const CardListFetcher = ({
   fetcherProps: { lane, queueType, tier },
 }: CardListFetcherProps) => {
   const dispatch = useDispatch();
-  const { showBoundary, resetBoundary } = useErrorBoundary();
 
-  React.useEffect(() => {
-    const fetchCards = async () => {
-      resetBoundary();
+  const config = {
+    params: {
+      size: 12,
+      page: 0,
+      position: lane,
+      type: queueType,
+      tier,
+    },
+  };
 
-      await defaultAxios
-        .get('/api/lol/boards', {
-          params: { size: 12, page: 0, position: lane, type: queueType, tier },
-        })
-        .then((response) => {
-          dispatch(cardActions.SET_CARDS(response.data.content));
-        })
-        .catch((error) => {
-          showBoundary(error);
-        });
-    };
-
-    fetchCards();
-  }, [lane, queueType, tier]);
+  const cardList: any = useGetData('/api/lol/boards', config);
+  dispatch(cardActions.SET_CARDS(cardList?.content));
 
   return <div>{children}</div>;
 };
